requirements.txt
requests
openai
python-dotenv
dateparser
rich
pytest
